% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df.flys4}
\alias{df.flys4}
\title{Stationary water levels from the FLYS 3-database}
\format{A \code{data.frame} with 4 columns: 
\describe{
  \item{river}{name of the relevant water body (type \code{character}).}
  \item{name}{of the FLYS 3 water level (type \code{character}). See details for more information.}
  \item{station}{rivers stationing (type \code{numeric}).}
  \item{w}{water level (cm above gauging station null, type \code{numeric}).} 
}}
\usage{
df.flys4
}
\description{
This dataset contains the 30 stationary 1D water levels for the
  rivers \strong{Elbe} and \strong{Rhein} originally stored in the 
  \href{https://www.bafg.de/DE/08_Ref/M2/03_Fliessgewmod/01_FLYS/flys_node.html}{FLYS3}-database.
}
\details{
The \code{name}ing of the water levels is \code{river}-specific.
}
\references{
Busch, N., Hammer, M., 2009: Einheitliche Grundlage für die Festlegung der 
  Bemessungswasserspiegellagen der Elbe auf der frei fließenden Strecke in 
  Deutschland. BfG-Bericht 1650, Bundesanstalt für Gewässerkunde, Koblenz, 
  Germany. DOI: \doi{10.5675/bfg-1650}
  
  HKV Hydrokontor, 2014: Erstellung eines SOBEK-River Modells für den Rhein 
  von Iffezheim bis Pannerdense Kop als Weiterentwicklung bestehender 
  SOBEK-RE Modelle. 
  
  Bundesanstalt für Gewässerkunde, 2016: FLYS – Flusshydrologischer 
  Webdienst. URL: \url{http://www.bafg.de/DE/08_Ref/M2/03_Fliessgewmod/01_FLYS/flys_node.html}
  
  DELTARES, 2018: SOBEK. URL: \url{https://www.deltares.nl/en/software/sobek}
}
\keyword{datasets}
